changed api auth checks, fix main page, users page, praise list
